controller: Fix extracting port names for setting queues
As the list representation of a dictionary does not always give a
logically ordered result, the internal interface of the switch sometimes
make it into other places than the last. As it is the shortest interface
name in the pool, sorting the list puts it into a deterministic place.

Relevant: #1