controller: Integrate configurability through YAML
This patch only extracts the most obvious parameters from the code and
some of them still have default values hardwired.
The long-term goal to eliminate class scope variables is also kept in
mind here (AdaptingMonitor13.configure being object-scope), so this part
should be less of a problem when carrying this elimination out.

Relevant: #4