controller: Make global QoS settings based on local measurements
Setting QoS for a virtual _network_ does not necessarily make sense if
we do it on a per-switch level as the switches for the same virtual
network may still have different settings. On the other hand, measuring
matching traffic on each switch provides a fine-grained enough picture
about the loads. Therefore, this patch applies global scope for QoS
manager, but keeps the local scope for logging and measurement.

Related: #26