controller: Clear queues from switches on adapting_monitor close
On the ryuapp.close() function, deleting the rules is necessary so that
leftovers do not disrupt the network further.  Due to rule deletion
problems in this stage, I finally decided to implement OpenFlow entry
deletion in a separate application which has a broader scope of deletion
than what I need, but I don't see any situations where it would be a
problem.

Along with this solution, queue deletion does not seem to be necessary.

Relevant: #13