controller: Handle exceptions due to long queue setting
As queue setting may easily last 10s of seconds it's really probable
that multiple setting requests are queued up. It turns out that waiting
requests get interrupted by new ones in specific cases which result in
an exception. To avoid long non-obvious stacktraces it's worth handling
it.

The rest of the calls do not seem to need this caution, so I leave them
alone for now.

Related: #29