controller: Implement a bit more thread awareness
To not block the monitoring with slow adaptation, I reorganized the
adaptation part so that it runs in a separate thread and in a way that
it doesn't cause a problem. Doing this I used semaphores at places where
it was absolutely necessary.

Another problem was a still not fully understood timeout done by
eventlet. Finally I figured out how to modify that number in an
acceptable manner. This is barely a solution, but I have spent enough
time with it for now...

Related: #29