controller: Make measurement records be optionally machine-readable
CSV formatting of the measurement output is now selectable in the config.

Headers keep being printed in every round of log writing, because the
current group of info that can make it into the experiment records would
mess it up anyway and filtering the unnecessary files are not too hard
anyway. So it's just a better compromise between machine and human
readability. Once those outputs that are now mixed in the records get
separated, headers may also become unnecessary to print every time.

Related: #25