controller: Retry too early queue setting once
Some time needs to elapse after ovsdb setting to set queues. The time
required for setting 2 to 3 rules are usually enough, but othervise,
queue setting simply fails. This patch does exactly one retry in case it
fails with this specific error.

Note: I could have used urrlib.Retry with its corresponding practice,
but this single retry did not seem to be worth it. If things get more
complicated, it should be considered for use.