controller: Make record logging periodical instead of event-based
I just realised that logging does not directly depend on information
provided by events. More importantly, though, to be able to merge
graphs from different sources (the controller and mininet), it's
necessary to put them on the same timeline with the same resolution.
Therefore it helps to have records for every second rather than every n
seconds. It's still not perfect due to the imperfection of greenthreads,
but it's still much easier to extrapolate.

Related: #25